requirements.txt
python-dotenv
langchain-core
langchain-google-genai
langchain
